+++ src/pages/Main.tsx
@@ -2,7 +2,7 @@ import { Button } from "@/components/ui/button";
 import type React from "react";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import ManualDialog from "@/components/ui/ManualDialog";
+import ManualDialog from "@/components/ui/dialog/ManualDialog";
 
 const MainPage: React.FC = () => {
 	const navigate = useNavigate();
